% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihs.R
\name{ihs}
\alias{ihs}
\title{Computes iHS Statistics (standardized IHH)}
\usage{
ihs(snp.list = "all", filter = 2, annot = T, write.xls = "both", plot = T)
}
\arguments{
\item{snp.list}{Character vector. It could be a scalar named "all" (default) for loading all SNPs
present in 'snps' directory, or a character vector with the exact names from the target-SNP files
(also supposed to be in 'populations' directory).}

\item{filter}{Numeric value for filtering significant iHS scores (default = 2).}

\item{annot}{Logical scalar. If TRUE (default), SNP annotation in xls files will be written with
correspondent SNP's genes, as well as ancestral and derived allele present in the dataset.
FALSE turns off this feature.}

\item{write.xls}{Scalar character. It could be set to "all.snps" for annotating all SNPs included
in this analysis, and "ss.snps" for annotating only SNPs statistically significant
(defined by 'filter' argument), or 'both' (default) for both annotation outputs.}

\item{plot}{Logical scalar. If TRUE (default), then distribution plots are generated and saved
as png figures.}
}
\value{
iHS files (csv or xls with SNP annotation) and iHS plots.
}
\description{
This function computes iHS Statistics (standardized IHH) for all target populations and SNPs.
}
\details{
All populations included in 'scanhh.RData' will be analyzed. This function creates a new directory in
'rehh_out' folder, named 'ihs', with two sub-directories: 'ihs/graphics' for plots, and 'ihs/tables'
for csv and xls files.

The default value for 'filter' parameter has been set to 2 accordingly the standard value described on
original paper. Two simple files (.csv) with iHS results will be generated for all included populations:
one with all SNPs included on analysis, and another with only statistically significant SNPs.

The 'annot' parameter turns on or off annotation procedure (default is TRUE, meaning 'on'), while argument
passed in 'write.xls' parameter tells which SNPs must be included on analysis ("all.snps", "ss.snps"
or "both"). Plot argument enables the generation of two plots/figures per population.

Except by annotation argument ('annot = T'), this function usually runs very fast. However,
depending on how many target-SNPs are present in the dataset, the annotation velocity may vary considerably.
}
\examples{
\dontrun{ihs()}

\dontrun{
ihs(snp.list = 'all', filter = 2, annot = T, write.xls = 'ss.snps', plot = F)
}

}
\seealso{
\code{\link{make.scanhh}}
}
