% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.binary.haplotypes.R
\name{make.binary.haplotypes}
\alias{make.binary.haplotypes}
\title{Make binary file with haplotypes}
\usage{
make.binary.haplotypes()
}
\arguments{
\item{chrs}{Numeric vector with chromosome ID(s). Default set to 1:22.}
}
\value{
R binary file named 'haps-sample.RData'.
}
\description{
This function imports all haplotypes present in the 'haplotypes' directory,
creating a single binary file for loading them all (as requested on subsequent functions).
}
\details{
Haplotype file names must start with "chr", followed by the chromosome ID
(represented as Arabic numerals), then ending with either ".haps" or ".sample"
extensions. Anything between the "chr[digit]" and the file extension can be written
for population identification.

\preformatted{Examples: chr1<ANYTHING>.haps and chr1<ANYTHING>.sample
          chr2.foo-bar.haps and chr2.foo-bar.sample
          chr3_foobar.haps and chr3_foobar.sample
          chr4.haps and chr4.sample
          etc
}
}
\examples{
\dontrun{
make.binary.haplotypes() # analyzing all chromosomes (default = 1:22)
make.binary.haplotypes(chrs = 1:10) # analyzing chromosomes from 1 to 10
make.binary.haplotypes(chrs = c(2,9,10,15,21)) # analyzing specific chromosomes
}

}
\seealso{
\code{\link{ldstart}}
}
\author{
Cainã Max Couto-Silva
}
