% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.scanhh.R
\name{make.scanhh}
\alias{make.scanhh}
\title{Create R binary file with EHH-based Statistics}
\usage{
make.input.files(populations = "all", chrs = 1:22, haps.subset = F, output)
}
\arguments{
\item{populations}{Character vector with target population's file names (assumed to be in
'populations' directory). If populations = 'all' (default), then all files (populations)
will be included on analysis.}

\item{chrs}{Numeric vector with chromosome ID(s).}

\item{output}{Character vector containing desired population names as output.}
}
\value{
R binary file named 'scanhh.RData'.
}
\description{
This function computes EHH-based Statistics which will be used in subsequent functions,
resulting in an R binary file containing all populations/chromosomes data analyzed.
}
\details{
The files acquired after running make.input.files() function will be used at this function.
Therefore, they are absolutely necessary (.inp and .thap files in 'rehh_in' directory).

By default, all populations files in 'populations' directory are included in this function,
although users can also specify the populations in a vector with the exact file names
(without providing the path, which is supposed to be in 'populations' directory).

Be careful when using 'output' parameter while using 'populations' parameter's default argument,
once you would not have control of the file names' order. If using the 'output' parameter,
you'd rather specify the populations' names as well. Default values to 'output' parameter
are automatically set according to acquired/loaded population's file names so that
there are no issues to worry about.
}
\examples{
\dontrun{make.input.files()}

\dontrun{
make.input.files(populations = c('Afr.txt', 'EUR.txt'),
                 chrs = 1:10, haps.subset = T,
                 output = c('Africans','Europeans'))
}

}
\seealso{
\code{\link{make.input.files}}
}
