% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_anc_alleles.R
\name{set_anc_alleles}
\alias{set_anc_alleles}
\title{Adjust Ancentral/Derived Alleles According to Reference}
\usage{
set_anc_alleles(sfile, ref, out)
}
\arguments{
\item{sfile}{Scalar character with the name of the dataset file on Shapeit format without extension (.haps/.sample).}

\item{ref}{Scalar character with the name of reference file (obtained by biomart R Package).}

\item{out}{Scalar character with desired name for the output file (without extension).}
}
\value{
Dataset with adjusted ancestral/derived alleles (.haps and .sample files).
}
\description{
This function uses a reference file obtained from biomart R Package, setting up ancestral alleles,
and excluding SNPs without ancestral information from dataset.
}
\details{
To obtain the Ensembl annotation for ancestral/derived SNPs, please check the function avaialable at:
https://github.com/cmcouto-silva/scripts/blob/master/ensembl_anc_allele.R
}
\examples{
\dontrun{
haps <- "dataset/mydataset_phased"
ref <- "reference_panels/ensembl_annot.txt"
out <- "out_folder/newfile"

set_anc_alleles(sfile, ref, out)
}
}
\author{
Cainã Max Couto-Silva
}
