% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpehh.R
\name{xpehh}
\alias{xpehh}
\title{Computes XP-EHH Statistics (standardized cross-population EHH)}
\usage{
xpehh(pop1, pop2, popname1, popname2, snp.list = "all", filter = 2, method = "both",
annot = T, write.xls = "both", plot = T)
}
\arguments{
\item{pop1}{Scalar character with the name of the first population to be compared to the second.}

\item{pop2}{Scalar character with the name of the second population to be compared to the first one.}

\item{popname1}{Scalar character with an optional name for the first population.}

\item{popname2}{Scalar character with an optional name for the second population.}

\item{snp.list}{Character vector. It could be a scalar named "all" (default) for loading all SNPs
present in 'snps' directory, or a character vector with the exact names from the target-SNP files
(also supposed to be in 'populations' directory).}

\item{filter}{Numeric value for filtering significant XP-EHH scores (default = 2).}

\item{method}{Scalar character. Indicates if applied method must be "bilateral", "unilateral",
or "both" (default).}

\item{annot}{Logical scalar. If TRUE (default), SNP annotation in xls files will be written with
correspondent SNP's genes, as well as ancestral and derived allele present in the dataset.
FALSE turns off this feature.}

\item{write.xls}{Scalar character. It could be set to "all.snps" for annotating all SNPs included
in this analysis, and "ss.snps" for annotating only SNPs statistically significant
(defined by 'filter' argument), or 'both' (default) for both annotation outputs.}

\item{plot}{Logical scalar. If TRUE (default), then XP-EHH-related plots are generated and saved
as png figures.}
}
\value{
XP-EHH files: simple result files (.csv), SNP annotation (.xls) and/or iHS plots (.png).
}
\description{
This function computes XP-EHH Statistics (standardized cross-population EHH) using two populations
and target-SNPs, then performs the SNPs' annotation and generate appropriate tables and graphics.
}
\details{
R binary file named 'scanhh.RData' must be loaded before running this function. Populations going
to be analyzed must be provided as 'scanhh_list$NAME', where 'NAME' is the population name (see @Examples
section). This function creates a new directory in 'rehh_out' folder, named 'xpehh', with two sub-directories:
'xpehh/graphics' for the plots, and 'ihs/tables' for csv and xls files.

The 'method' parameter indicates which method will be applied ('bilateral', 'unilateral', or 'both';
default = 'both'). The default value for 'filter' parameter has been set to 2 accordingly the
standard value described on original paper.

The 'annot' parameter turns on or off annotation procedure (default is TRUE, meaning 'on'), while argument
passed in 'write.xls' parameter tells which SNPs must be included on analysis ("all.snps", "ss.snps"
or "both"). Plot argument enables the generation of two plots/figures per population.

If all parameter's arguments are used as they are by default, then a total of four simple files (.csv)
containing XP-EHH results will be generated by analysis to both methods and both 'write.xls' arguments.

Except by annotation argument ('annot = T'), this function usually runs very fast. However,
depending on how many target-SNPs are present in the dataset, the annotation velocity may vary considerably.
}
\examples{
\dontrun{
load('scanhh.RData')
xpehh(pop1 = scanhh_list$Africans, pop2 = scanhh_list$Europeans)}

\dontrun{
load('scanhh.RData')
xpehh(pop1 = scanhh_list$Africans, pop2 = scanhh_list$Europeans), snp.list = 'all', filter = 2,
method = 'unilateral', annot = T, write.xls = 'ss.snps', plot = T)
}

}
\seealso{
\code{\link{make.scanhh}}
}
